+++ macros/src/main/scala-3/com/softwaremill/macwire/internals/utils.scala
@@ -1,0 +1,6 @@
+package com.softwaremill.macwire.internals
+
+import scala.quoted.*
+
+def showTypeName[T: Type](using q: Quotes): String = showTypeName(using q)(q.reflect.TypeRepr.of[T])
+def showTypeName(using q: Quotes)(t: q.reflect.TypeRepr): String = t.typeSymbol.name
