+++ tests/src/test/scala-3/com/softwaremill/macwire/AutowireCompileTests.scala
@@ -1,0 +1,10 @@
+package com.softwaremill.macwire
+
+class AutowireCompileTests extends CompileTestsSupport:
+  override protected val testsSubdirectory = Some("autowire")
+
+  runTestsWith(
+    expectedFailures = List(
+      "privateConstructor" -> List("Cannot find a dependency or a public constructor for: A, while wiring: A")
+    )
+  )
